Add auth service

+++ tasks-service/src/services/taskCrudService.ts
@@ -6,16 +6,29 @@ import {
 } from "../types/tasks/Task.js";
 
 class TaskCrudService {
-  async getAll(): Promise<Task[]> {
-    return prisma.task.findMany();
+  async getAll(userId: string): Promise<Task[]> {
+    return prisma.task.findMany({
+      where: {
+        userId,
+      },
+    });
   }
 
-  async getById(id: string): Promise<Task | null> {
-    return prisma.task.findFirst({ where: { id } });
+  async getById(id: string, userId: string): Promise<Task | null> {
+    return prisma.task.findFirst({ where: { id, userId } });
   }
 
-  async deleteById(id: string): Promise<Task | null> {
+  async deleteById(id: string, userId: string): Promise<Task | null> {
     try {
+      const task = await prisma.task.findFirst({
+        where: {
+          id,
+          userId,
+        },
+      });
+      if (!task) {
+        return task;
+      }
       return await prisma.task.delete({
         where: {
           id,
@@ -26,15 +39,25 @@ class TaskCrudService {
     }
   }
 
-  async createOne(task: TaskCreatePayload): Promise<Task> {
-    return prisma.task.create({ data: task });
+  async createOne(task: TaskCreatePayload, userId: string): Promise<Task> {
+    return prisma.task.create({ data: { ...task, userId } });
   }
 
   async updateOne(
     id: string,
+    userId: string,
     payload: TaskUpdatePayload
   ): Promise<Task | null> {
     try {
+      const task = await prisma.task.findFirst({
+        where: {
+          id,
+          userId,
+        },
+      });
+      if (!task) {
+        return task;
+      }
       return await prisma.task.update({
         data: payload,
         where: {
